# Escape zero-width characters