--- a/cmd/gitprompt/main.go
+++ b/cmd/gitprompt/main.go
@@ -2,42 +2,99 @@ package main
 
 import (
 	"bufio"
+	"bytes"
 	"context"
 	"flag"
 	"fmt"
+	"io"
 	"log"
 	"os"
 	"os/exec"
+	"strconv"
 	"strings"
 	"time"
 
 	"github.com/abhinav/gitprompt"
-	"github.com/fatih/color"
 )
 
 func main() {
-	color.NoColor = false
 	log.SetFlags(0)
-	if err := run(os.Args[1:]); err != nil {
+	if err := run(os.Args[1:], os.Stdout); err != nil {
 		log.Fatal(err)
 	}
 }
 
+type color []byte
+
 var (
-	branchColor   = color.New(color.FgMagenta, color.Bold)
-	stagedColor   = color.New(color.FgRed)
-	conflictColor = color.New(color.FgRed)
-	changedColor  = color.New(color.FgBlue)
-	cleanColor    = color.New(color.FgGreen)
+	_branchColor   = color("\x1b[35;1m") // magenta; bold
+	_stagedColor   = color("\x1b[31m")   // red
+	_conflictColor = _stagedColor        // red
+	_changedColor  = color("\x1b[34m")   // blue
+	_cleanColor    = color("\x1b[32m")   // green
+
+	_reset = color("\x1b[0m")
 )
 
-func run(args []string) error {
+type shell struct {
+	// Byte sequences to write before and after each zero-width sequence
+	// of characters in the prompt.
+	ZWOpen, ZWClose []byte
+}
+
+var (
+	_zsh  = shell{ZWOpen: []byte("%{"), ZWClose: []byte("%}")}
+	_bash = shell{ZWOpen: []byte{0x01}, ZWClose: []byte{0x02}}
+)
+
+type promptBuffer struct {
+	bytes.Buffer
+
+	Shell shell
+
+	// Used for AppendInt.
+	intbuff [4]byte
+}
+
+func (w *promptBuffer) Itoa(i int64) {
+	w.Write(strconv.AppendInt(w.intbuff[:0], i, 10))
+}
+
+func (w *promptBuffer) Color(c color) {
+	w.Write(w.Shell.ZWOpen)
+	w.Write(c)
+	w.Write(w.Shell.ZWClose)
+}
+
+func (w *promptBuffer) Reset() {
+	w.Color(_reset)
+}
+
+const (
+	_check    = '✔'
+	_cross    = '✖'
+	_dot      = '●'
+	_dots     = '…'
+	_down     = '↓'
+	_parClose = ')'
+	_parOpen  = '('
+	_pipe     = '|'
+	_plus     = '✚'
+	_up       = '↑'
+)
+
+func run(args []string, stdout io.Writer) error {
 	var (
 		timeout  time.Duration
 		noStatus bool
 	)
 
 	flag := flag.NewFlagSet("gitprompt", flag.ContinueOnError)
+	flag.Usage = func() {
+		fmt.Fprintf(flag.Output(), "usage: %v shell\n", flag.Name())
+		flag.PrintDefaults()
+	}
+
 	flag.DurationVar(&timeout, "timeout", 0,
 		"amount of time the 'git status' command is allowed to take; unlimited if 0")
 
@@ -46,6 +103,21 @@ func run(args []string) error {
 
 	if err := flag.Parse(args); err != nil {
 		return err
+	}
+
+	if flag.NArg() == 0 {
+		flag.Usage()
+		return nil
+	}
+
+	var shell shell
+	switch flag.Arg(0) {
+	case "bash":
+		shell = _bash
+	case "zsh":
+		shell = _zsh
+	default:
+		return fmt.Errorf("unsupported shell %q", flag.Arg(0))
 	}
 
 	ctx := context.Background()
@@ -71,44 +143,69 @@ func run(args []string) error {
 		return nil // not a git repo
 	}
 
-	fmt.Printf("(%v", branchColor.Sprint(s.Branch))
-	defer fmt.Print(")")
+	w := promptBuffer{Shell: shell}
+	buildPrompt(&w, s)
+	w.WriteTo(stdout)
+	return nil
+}
+
+func buildPrompt(w *promptBuffer, s *Status) {
+	w.WriteRune(_parOpen)
+	defer w.WriteRune(_parClose)
+
+	w.Color(_branchColor)
+	w.WriteString(s.Branch)
+	w.Reset()
 
 	if s.BranchOnly {
-		return nil
+		return
 	}
 
 	if s.Behind > 0 {
-		fmt.Printf("↓%d", s.Behind)
+		w.WriteRune(_down)
+		w.Itoa(s.Behind)
 	}
+
 	if s.Ahead > 0 {
-		fmt.Printf("↑%d", s.Ahead)
+		w.WriteRune(_up)
+		w.Itoa(s.Ahead)
 	}
-	fmt.Printf("|")
+
+	w.WriteRune(_pipe)
 
 	if s.Staged > 0 {
-		stagedColor.Printf("●%d", s.Staged)
+		w.Color(_stagedColor)
+		w.WriteRune(_dot)
+		w.Itoa(s.Staged)
+		w.Reset()
 	}
 
 	if s.Conflicts > 0 {
-		conflictColor.Printf("✖%d", s.Conflicts)
+		w.Color(_conflictColor)
+		w.WriteRune(_cross)
+		w.Itoa(s.Conflicts)
+		w.Reset()
 	}
 
 	if s.Changed > 0 {
-		changedColor.Printf("✚%d", s.Changed)
+		w.Color(_changedColor)
+		w.WriteRune(_plus)
+		w.Itoa(s.Changed)
+		w.Reset()
 	}
 
 	if s.Untracked > 0 {
-		fmt.Printf("…%d", s.Untracked)
+		w.WriteRune(_dots)
+		w.Itoa(s.Untracked)
 	}
 
 	switch {
 	case s.Changed > 0, s.Conflicts > 0, s.Staged > 0, s.Untracked > 0:
 	default:
-		cleanColor.Printf("✔")
+		w.Color(_cleanColor)
+		w.WriteRune(_check)
+		w.Reset()
 	}
-
-	return nil
 }
 
 func gitStatus(ctx context.Context) (*Status, error) {