--- a/go.mod
+++ b/go.mod
@@ -3,7 +3,6 @@ module github.com/abhinav/gitprompt
 go 1.12
 
 require (
-	github.com/fatih/color v1.5.0
 	github.com/tcnksm/ghr v0.13.0
 	golang.org/x/lint v0.0.0-20200302205851-738671d3881b
 	honnef.co/go/tools v0.0.1-2020.1.3