If gitprompt is used in a single-line prompt, it makes tab completion
behave poorly. For example, when the prompt says,

    (master|✔) $ ls |

Pressing tab at this point will fire the tab completion behavior, but
it will then indent the ls a few characters, leaving a copy of the old
one in place.

    (master|✔) $ ls                  ls |

This is because the position of the cursor is calculated based on the
length of the propmpt text. Zsh and Bash don't realize that the escape
codes in the output do not add to the length of the text.

Zsh supports `%{...%}` in prompts which treats the contents inside the
braces as a zero-width escape sequence.
See also http://zsh.sourceforge.net/Doc/Release/Prompt-Expansion.html#Visual-effects.

Similarly, Bash supports `\[...\]` for a similar purpose, however that's
just short for `\001` and `\002`.
See also https://stackoverflow.com/a/32267141.

This changes gitprompt to support both forms of escapes to make
gitprompt play nicely with tab completion on single-line prompts.
In the process, this drops the dependency on fatih/color, and instead
manages colors locally.

And while at it, this also speeds up the command by using more efficient
operations to build the prompt components.

Before:

```
$ hyperfine './gitprompt zsh'
Benchmark #1: ./gitprompt zsh
  Time (mean ± σ):      91.6 ms ±   2.7 ms    [User: 9.7 ms, System: 77.9 ms]
  Range (min … max):    88.7 ms …  99.2 ms    31 runs
```

After:

```
$ hyperfine './gitprompt zsh'
Benchmark #1: ./gitprompt zsh
  Time (mean ± σ):      49.3 ms ±   2.2 ms    [User: 6.0 ms, System: 40.5 ms]
  Range (min … max):    45.2 ms …  55.3 ms    55 runs
```